refactor(loader): adjusted tests to the new namespace/structure
Refs [#15489]
